Move computeSigningRoot, computeForkDigest, computeDomain to MiscHelpers
