+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/common/util/BeaconStateUtil.java
@@ -20,15 +20,12 @@ import java.util.List;
 import java.util.stream.Stream;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.config.SpecConfig;
 import tech.pegasys.teku.spec.datastructures.blocks.BeaconBlock;
 import tech.pegasys.teku.spec.datastructures.operations.Attestation;
 import tech.pegasys.teku.spec.datastructures.state.Fork;
-import tech.pegasys.teku.spec.datastructures.state.ForkData;
-import tech.pegasys.teku.spec.datastructures.state.SigningData;
 import tech.pegasys.teku.spec.datastructures.state.Validator;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconState;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconStateCache;
@@ -36,11 +33,8 @@ import tech.pegasys.teku.spec.logic.common.helpers.BeaconStateAccessors;
 import tech.pegasys.teku.spec.logic.common.helpers.MiscHelpers;
 import tech.pegasys.teku.spec.logic.common.helpers.Predicates;
 import tech.pegasys.teku.spec.schemas.SchemaDefinitions;
-import tech.pegasys.teku.ssz.Merkleizable;
 import tech.pegasys.teku.ssz.SszList;
 import tech.pegasys.teku.ssz.collections.SszBitvector;
-import tech.pegasys.teku.ssz.collections.SszByteVector;
-import tech.pegasys.teku.ssz.primitive.SszUInt64;
 import tech.pegasys.teku.ssz.type.Bytes4;
 
 @SuppressWarnings("unused")
@@ -88,20 +82,12 @@ public class BeaconStateUtil {
         : currentEpoch.plus(1);
   }
 
-  public Bytes32 computeDomain(Bytes4 domainType) {
-    return computeDomain(domainType, specConfig.getGenesisForkVersion(), Bytes32.ZERO);
-  }
-
   public Bytes32 getPreviousDutyDependentRoot(BeaconState state) {
     return getDutyDependentRoot(state, beaconStateAccessors.getPreviousEpoch(state));
   }
 
   public Bytes32 getCurrentDutyDependentRoot(BeaconState state) {
     return getDutyDependentRoot(state, beaconStateAccessors.getCurrentEpoch(state));
-  }
-
-  public Bytes4 computeForkDigest(Bytes4 currentVersion, Bytes32 genesisValidatorsRoot) {
-    return new Bytes4(computeForkDataRoot(currentVersion, genesisValidatorsRoot).slice(0, 4));
   }
 
   public Bytes32 getDomain(BeaconState state, Bytes4 domainType, UInt64 messageEpoch) {
@@ -123,7 +109,7 @@ public class BeaconStateUtil {
         (epoch.compareTo(fork.getEpoch()) < 0)
             ? fork.getPrevious_version()
             : fork.getCurrent_version();
-    return computeDomain(domainType, forkVersion, genesisValidatorsRoot);
+    return miscHelpers.computeDomain(domainType, forkVersion, genesisValidatorsRoot);
   }
 
   public List<UInt64> getEffectiveBalances(final BeaconState state) {
@@ -139,15 +125,6 @@ public class BeaconStateUtil {
                                 ? validator.getEffective_balance()
                                 : UInt64.ZERO)
                     .collect(toUnmodifiableList()));
-  }
-
-  public Bytes computeSigningRoot(Merkleizable object, Bytes32 domain) {
-    return new SigningData(object.hashTreeRoot(), domain).hashTreeRoot();
-  }
-
-  public Bytes computeSigningRoot(UInt64 number, Bytes32 domain) {
-    SigningData domainWrappedObject = new SigningData(SszUInt64.of(number).hashTreeRoot(), domain);
-    return domainWrappedObject.hashTreeRoot();
   }
 
   public boolean all(SszBitvector bitvector, int start, int end) {
@@ -182,12 +159,6 @@ public class BeaconStateUtil {
         .map(validatorIndex -> state.getValidators().get(validatorIndex).getEffective_balance());
   }
 
-  public Bytes32 computeSigningRoot(Bytes bytes, Bytes32 domain) {
-    SigningData domainWrappedObject =
-        new SigningData(SszByteVector.computeHashTreeRoot(bytes), domain);
-    return domainWrappedObject.hashTreeRoot();
-  }
-
   public int computeSubnetForAttestation(final BeaconState state, final Attestation attestation) {
     final UInt64 attestationSlot = attestation.getData().getSlot();
     final UInt64 committeeIndex = attestation.getData().getIndex();
@@ -208,20 +179,6 @@ public class BeaconStateUtil {
         committeeIndex,
         beaconStateAccessors.getCommitteeCountPerSlot(
             state, miscHelpers.computeEpochAtSlot(attestationSlot)));
-  }
-
-  private Bytes32 computeDomain(
-      Bytes4 domainType, Bytes4 forkVersion, Bytes32 genesisValidatorsRoot) {
-    final Bytes32 forkDataRoot = computeForkDataRoot(forkVersion, genesisValidatorsRoot);
-    return computeDomain(domainType, forkDataRoot);
-  }
-
-  private Bytes32 computeDomain(final Bytes4 domainType, final Bytes32 forkDataRoot) {
-    return Bytes32.wrap(Bytes.concatenate(domainType.getWrappedBytes(), forkDataRoot.slice(0, 28)));
-  }
-
-  private Bytes32 computeForkDataRoot(Bytes4 currentVersion, Bytes32 genesisValidatorsRoot) {
-    return new ForkData(currentVersion, genesisValidatorsRoot).hashTreeRoot();
   }
 
   private Bytes32 getDutyDependentRoot(final BeaconState state, final UInt64 epoch) {
